Section 5: Authorization - Task Ownership

+++ src/tasks/task.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity'
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TaskService {
@@ -13,12 +14,18 @@ export class TaskService {
         private taskRepository: TaskRepository
     ) { }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto)
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user)
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(
+        id: number,
+        user: User
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } })
 
         if (!found) {
             throw new NotFoundException('Task could not be found')
@@ -27,19 +34,29 @@ export class TaskService {
         return found
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const found = await this.getTaskById(id)
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const found = await this.getTaskById(id, user)
         found.status = status
         await found.save()
         return found
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(
+        id: number,
+        user: User
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id })
 
         if (!result.affected) {
             throw new NotFoundException('Task could not be deleted')
